+++ CHANGELOG.md
@@ -9,11 +9,16 @@
   - `sort_array`
   - `array_min`
   - `array_max`
+- Added support for async execution of multi-query dataframe containing binding variables.
 - Added support for renaming multiple columns in `DataFrame.rename`.
 
 ### Improvements
 
 - Redundant dependency `typing-extensions` removed.
+
+### Dependency updates
+
+- Updated ``snowflake-connector-python`` to 3.0.4.
 
 ## 1.5.1 (2023-06-20)